Let Message component use Redux state
