+++ decadegraphy/static/js/Campaign/components/Message.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import Helpers from '../../helpers.js'
-
-import SignUp from './SignUp'
+import dispatch from 'store'
+import * as events from 'events.js'
+import { connect } from 'react-redux'
 
 class Message extends React.Component {
   constructor (props) {
     super(props)
     this.state = {
-      roles: [],
       qrCodes: []
     }
   }
@@ -34,8 +34,11 @@ class Message extends React.Component {
           })
 
           if (data) {
+            dispatch({
+              type: events.SIGNUP_ROLES_FETCHED,
+              payload: { roles: data.roles }
+            })
             this.setState({
-              roles: data.roles.split(','),
               qrCodes
             })
           }
@@ -48,13 +51,14 @@ class Message extends React.Component {
   }
 
   render () {
-    let roleNames = SignUp.defaultProps.roleNames
+    const { roles } = this.props
+
     return (
       <div className="enroll-success">
         <img src="" alt="" className="success-icon" />
         <div className="content">
           <p>报名成功</p>
-          <p>您的身份是：{this.state.roles.map(r => roleNames[r - 1]).join('，')}</p>
+          <p>您的身份是：{Helpers.translateRoleNames(roles)}</p>
           <p>1. 进群前请先阅读<a href="#">活动流程指引与约拍须知</a></p>
           <p>2. 扫描下方二维码加入城市拍摄微信群，开始你的Decadegraphy旅程</p>
           <ul className="qrcode">
@@ -66,4 +70,12 @@ class Message extends React.Component {
   }
 }
 
-export default Message
+function mapStateToProps (state, props) {
+  const { roles } = state.campaigns
+
+  return {
+    roles
+  }
+}
+
+export default connect(mapStateToProps)(Message)
