+++ decadegraphy/static/js/events.js
@@ -1,3 +1,4 @@
 export const SIGNUP_TOGGLE_ROLE = 'SIGNUP_TOGGLE_ROLE'
 export const CHANGE_STEP = 'CHANGE_STEP'
 export const TWITTER_ID_FETCHED = 'TWITTER_ID_FETCHED'
+export const SIGNUP_ROLES_FETCHED = 'SIGNUP_ROLES_FETCHED'
